Moving hard delete clases out of PersistentIndex
Moving hard delete functionality out of PersistentIndex because of the logical seperation
Also helps reason about log segmentation and compaction changes down the line
